Trait pallet_asset_management::pallet::Config

pub trait Config: Config + Config + Config + Config + Config + Config + Config + Config + Config + Config {

    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Call: Parameter + UnfilteredDispatchable<Origin = <Self as Config>::Origin> + From<Call<Self>> + GetDispatchInfo;
    type Delay: Get<Self::BlockNumber>;
    type CheckDelay: Get<Self::BlockNumber>;
    type InvestorVoteAmount: Get<u128>;
    type Currency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>;
    type WeightInfo: WeightInfo;
    type Guaranty: Get<u32>;
    type RoR: Get<Percent>;
    type MinimumDepositVote: Get<BalanceOf<Self>>;
    type RepFees: Get<BalanceOf<Self>>;
    type ContractLength: Get<Self::BlockNumber>;
    type CheckPeriod: Get<Self::BlockNumber>;
    type RentCheck: Get<Self::BlockNumber>;
    type Lease: Get<u32>;
    type Maintenance: Get<Percent>;

}

Configure the pallet by specifying the parameters and types on which it depends.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

Because this pallet emits events, it depends on the runtime’s definition of an event.

type Call: Parameter + UnfilteredDispatchable<Origin = <Self as Config>::Origin> + From<Call<Self>> + GetDispatchInfo

type Delay: Get<Self::BlockNumber>

type CheckDelay: Get<Self::BlockNumber>

type InvestorVoteAmount: Get<u128>

type Currency: Currency<Self::AccountId> + ReservableCurrency<Self::AccountId>

type WeightInfo: WeightInfo

type Guaranty: Get<u32>

Number of months for the guaranty deposit

type RoR: Get<Percent>

Return on Rent

type MinimumDepositVote: Get<BalanceOf<Self>>

type RepFees: Get<BalanceOf<Self>>

Fees payed to the Representative by the tenant, to provide a judgement

type ContractLength: Get<Self::BlockNumber>

Lease period in number of blocks

type CheckPeriod: Get<Self::BlockNumber>

Period between check of Referendum status

type RentCheck: Get<Self::BlockNumber>

Period between check of rent payment status for active tenants

type Lease: Get<u32>

Lease period in number of months

type Maintenance: Get<Percent>

Maintenance fees percentage taken on monthly rent