Trait pallet_asset_management::PaymentHandler

pub trait PaymentHandler<T>where
    T: Config,{
    fn create_payment(
        from: &<T as Config>::AccountId,
        to: &<T as Config>::AccountId,
        amount: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance,
        payment_state: PaymentState<T>,
        incentive_percentage: Percent,
        remark: Option<&[u8]>
    ) -> Result<PaymentDetail<T>, DispatchError>;
    fn reserve_payment_amount(
        from: &<T as Config>::AccountId,
        to: &<T as Config>::AccountId,
        payment: PaymentDetail<T>
    ) -> Result<(), DispatchError>;
    fn settle_payment(
        from: &<T as Config>::AccountId,
        to: &<T as Config>::AccountId,
        recipient_share: Percent
    ) -> Result<(), DispatchError>;
    fn get_payment_details(
        from: &<T as Config>::AccountId,
        to: &<T as Config>::AccountId
    ) -> Option<PaymentDetail<T>>;
}

trait that defines how to create/release payments for users

Required Methods

fn create_payment(
from: &<T as Config>::AccountId,
to: &<T as Config>::AccountId,
amount: <<T as Config>::Currency as Currency<<T as Config>::AccountId>>::Balance,
payment_state: PaymentState<T>,
incentive_percentage: Percent,
remark: Option<&[u8]>
) -> Result<PaymentDetail<T>, DispatchError>

Create a PaymentDetail from the given payment details Calculate the fee amount and store PaymentDetail in storage Possible reasons for failure include:

Payment already exists and cannot be overwritten

fn reserve_payment_amount(
from: &<T as Config>::AccountId,
to: &<T as Config>::AccountId,
payment: PaymentDetail<T>
) -> Result<(), DispatchError>

Attempt to reserve the amount from the caller If not possible then return Error. Possible reasons for failure include:

User does not have enough balance.

fn settle_payment(
from: &<T as Config>::AccountId,
to: &<T as Config>::AccountId,
recipient_share: Percent
) -> Result<(), DispatchError>

The payment does not exist

The unreserve operation fails

The transfer operation fails

fn get_payment_details(
from: &<T as Config>::AccountId,
to: &<T as Config>::AccountId
) -> Option<PaymentDetail<T>>

Attempt to fetch the details of a payment from the given payment_id Possible reasons for failure include:

The payment does not exist