Trait pallet_asset_management::UnfilteredDispatchable

pub trait UnfilteredDispatchable {
    type Origin;

    fn dispatch_bypass_filter(
        self,
        origin: Self::Origin
    ) -> Result<PostDispatchInfo, DispatchErrorWithPostInfo<PostDispatchInfo>>;
}

Type that can be dispatched with an origin but without checking the origin filter.

Implemented for pallet dispatchable type by decl_module and for runtime dispatchable by construct_runtime.

Required Associated Types

type Origin

The origin type of the runtime, (i.e. frame_system::Config::Origin).

Required Methods

fn dispatch_bypass_filter(
self,
origin: Self::Origin
) -> Result<PostDispatchInfo, DispatchErrorWithPostInfo<PostDispatchInfo>>

Dispatch this call but do not check the filter in origin.